split up utils.py; Fix: incorrect naming (metadata -> raw txn data); use of dataclasses and TypedDict
