+++ xrpl/utils/txn_parser/__init__.py
@@ -1,17 +1,17 @@
 """Transaction parser."""
 
-from xrpl.utils.tx_parser.balance_changes import (
+from xrpl.utils.txn_parser.balance_changes import (
     parse_balance_changes,
     parse_final_balances,
     parse_previous_balances,
 )
-from xrpl.utils.tx_parser.orderbook_changes import parse_order_book_changes
-from xrpl.utils.tx_parser.utils import XRPLMetadataException
+from xrpl.utils.txn_parser.order_book_changes import parse_order_book_changes
+from xrpl.utils.txn_parser.utils import XRPLTxnFieldsException
 
 __all__ = [
     "parse_balance_changes",
     "parse_final_balances",
     "parse_previous_balances",
     "parse_order_book_changes",
-    "XRPLMetadataException",
+    "XRPLTxnFieldsException",
 ]
